Move lone `Unit`tests to the corresponding `Basic` test class [API-383]
