--- tests/Basic/WebArtistTest.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Tests\Basic;
-
-use App\Models\Web\Artist;
-
-class WebArtistTest extends BasicTestCase
-{
-    protected $model = Artist::class;
-
-    protected $route = 'web-artists';
-}
