--- tests/Basic/WebExhibitionTest.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Tests\Basic;
-
-use App\Models\Web\Exhibition;
-
-class WebExhibitionTest extends BasicTestCase
-{
-    protected $model = Exhibition::class;
-
-    protected $route = 'web-exhibitions';
-}
